pyproject.toml
[project]
name = "tnlcm"
version = "0.4.5"
description = "TNLCM has been designed as a modular application, with the intention of making certain components easily replaceable or extended, while minimizing the effect of changes in other parts of the application."
readme = "docs/README.md"
requires-python = ">=3.13"
authors = [{ name = "Carlos Andreo", email = "c.andreo@uma.es" }]

dependencies = [
  "Flask==3.1.0",
  "Flask-Cors==5.0.1",
  "flask-restx==1.3.0",
  "Flask-JWT-Extended==4.7.1",
  "gunicorn==23.0.0",
  "mongoengine==0.29.1",
  "python-dotenv==1.0.1",
  "python-jenkins==1.8.2",
  "ruamel.yaml==0.18.10",
  "Werkzeug==3.1.3",
]
